Move covers() to utils.py
